fix a couple of bugs and formatting
